See Changelog 2003-09-14 11:00 UTC-0500 Paul Tucker

+++ xharbour/contrib/hbzlib/zipplatform.cpp
@@ -1,7 +1,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 // $Workfile: ZipPlatform.cpp $
 // $Archive: /ZipArchive/ZipPlatform.cpp $
-// $Date: 2003/08/20 15:03:53 $ $Author: lculik $
+// $Date: 2003/08/20 19:33:40 $ $Author: lculik $
 ////////////////////////////////////////////////////////////////////////////////
 // This source file is part of the ZipArchive library source distribution and
 // is Copyright 2000-2003 by Tadeusz Dracz (http://www.artpol-software.com/)
@@ -10,7 +10,7 @@
 // modify it under the terms of the GNU General Public License
 // as published by the Free Software Foundation; either version 2
 // of the License, or (at your option) any later version.
-// 
+//
 // For the licensing details see the file License.txt
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -30,7 +30,7 @@
 
 #include <direct.h>
 #include <io.h>
-#include <time.h> 
+#include <time.h>
 #include "zippathcomponent.h"
 #include "zipcompatibility.h"
 
@@ -38,7 +38,7 @@ const TCHAR CZipPathComponent::m_cSeparator = _T('\\');
 
 
 #ifndef _UTIMBUF_DEFINED
-#define _utimbuf utimbuf 
+#define _utimbuf utimbuf
 #endif
 
 DWORD ZipPlatform::GetDeviceFreeSpace(LPCTSTR lpszPath)
@@ -60,7 +60,7 @@ DWORD ZipPlatform::GetDeviceFreeSpace(LPCTSTR lpszPath)
 			&BytesPerSector,
 			&NumberOfFreeClusters,
 			&TotalNumberOfClusters))
-		
+
 				return 0;
 	}
 	__int64 total = SectorsPerCluster * BytesPerSector * NumberOfFreeClusters;
@@ -69,7 +69,7 @@ DWORD ZipPlatform::GetDeviceFreeSpace(LPCTSTR lpszPath)
 
 bool ZipPlatform::GetFileSize(LPCTSTR lpszFileName, DWORD& dSize)
 {
-	HANDLE f = CreateFile(lpszFileName, GENERIC_READ, FILE_SHARE_READ, 
+	HANDLE f = CreateFile(lpszFileName, GENERIC_READ, FILE_SHARE_READ,
 		NULL, OPEN_EXISTING, 0, NULL);
 	if (!f)
 		return false;
@@ -98,7 +98,7 @@ CZipString ZipPlatform::GetTmpFileName(LPCTSTR lpszPath, DWORD iSizeNeeded)
 			DWORD size = GetTempPath(0, NULL);
 			if (size == 0)
 				return empty;
-		
+
 			GetTempPath(size, tempPath.GetBuffer(size));
 			tempPath.ReleaseBuffer();
 			if (GetDeviceFreeSpace(tempPath) < iSizeNeeded)
@@ -139,11 +139,11 @@ bool ZipPlatform::GetFileAttr(LPCTSTR lpFileName, DWORD& uAttr)
 {
 	// not using MFC due to MFC bug (attr is one byte there)
 	DWORD temp = ::GetFileAttributes(lpFileName);
-	if (temp == -1)
+	if (temp == ( DWORD )-1)
 		return false;
 	uAttr = temp;
 	return true;
-	
+
 }
 
 bool ZipPlatform::GetFileModTime(LPCTSTR lpFileName, time_t & ttime)
@@ -214,7 +214,7 @@ ZIPINLINE  bool ZipPlatform::RemoveFile(LPCTSTR lpszFileName, bool bThrow)
 	if (!::DeleteFile((LPTSTR)lpszFileName))
 		if (bThrow)
 			CZipException::Throw(CZipException::notRemoved, lpszFileName);
-		else 
+		else
 			return false;
 	return true;
 
@@ -224,7 +224,7 @@ ZIPINLINE  bool ZipPlatform::RenameFile( LPCTSTR lpszOldName, LPCTSTR lpszNewNam
 	if (!::MoveFile((LPTSTR)lpszOldName, (LPTSTR)lpszNewName))
 		if (bThrow)
 			CZipException::Throw(CZipException::notRenamed, lpszOldName);
-		else 
+		else
 			return false;
 	return true;
 
@@ -258,7 +258,7 @@ ZIPINLINE bool ZipPlatform::GetSystemCaseSensitivity()
 	return false;
 }
 
-#ifdef _UNICODE	
+#ifdef _UNICODE
 int ZipPlatform::WideToSingle(LPCTSTR lpWide, CZipAutoBuffer &szSingle)
 {
 	size_t wideLen = wcslen(lpWide);
@@ -269,12 +269,12 @@ int ZipPlatform::WideToSingle(LPCTSTR lpWide, CZipAutoBuffer &szSingle)
 	}
 
 	// iLen does not include terminating character
-	int iLen = WideCharToMultiByte(CP_ACP,0, lpWide, wideLen, szSingle, 
+	int iLen = WideCharToMultiByte(CP_ACP,0, lpWide, wideLen, szSingle,
 		0, NULL, NULL);
 	if (iLen > 0)
 	{
 		szSingle.Allocate(iLen, true);
-		iLen = WideCharToMultiByte(CP_ACP,0, lpWide , wideLen, szSingle, 
+		iLen = WideCharToMultiByte(CP_ACP,0, lpWide , wideLen, szSingle,
 			iLen, NULL, NULL);
 		ASSERT(iLen != 0);
 	}
@@ -293,7 +293,7 @@ int ZipPlatform::SingleToWide(const CZipAutoBuffer &szSingle, CZipString& szWide
 	int iLen = MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, szSingle.GetBuffer(), singleLen, NULL, 0);
 	if (iLen > 0)
 	{
-		iLen = MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, szSingle.GetBuffer(), singleLen, 
+		iLen = MultiByteToWideChar(CP_ACP, MB_PRECOMPOSED, szSingle.GetBuffer(), singleLen,
 			szWide.GetBuffer(iLen) , iLen);
 		szWide.ReleaseBuffer(iLen);
 		ASSERT(iLen != 0);
